Move some stuff around in prep for a VL2 rework and identity rework.
