+++ network-hypervisor/src/vl2/mod.rs
@@ -1,16 +1,16 @@
 // (c) 2020-2022 ZeroTier, Inc. -- currently proprietary pending actual release and licensing. See LICENSE.md.
 
+mod iproute;
 mod multicastgroup;
 mod networkid;
-mod revocation;
 mod switch;
+mod topology;
 
 pub mod multicastauthority;
-pub mod networkconfig;
 pub mod rule;
 pub mod v1;
 
+pub use iproute::IpRoute;
 pub use multicastgroup::MulticastGroup;
 pub use networkid::NetworkId;
-pub use revocation::Revocation;
 pub use switch::{Switch, SwitchInterface};
