+++ network-hypervisor/src/vl2/topology.rs
@@ -1,0 +1,3 @@
+pub struct Member {}
+
+pub struct Topology {}
