+++ network-hypervisor/src/vl2/v1/mod.rs
@@ -1,5 +1,7 @@
 mod certificateofmembership;
 mod certificateofownership;
+pub mod networkconfig;
+mod revocation;
 mod tag;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -15,4 +17,5 @@ pub enum CredentialType {
 
 pub use certificateofmembership::CertificateOfMembership;
 pub use certificateofownership::{CertificateOfOwnership, Thing};
+pub use revocation::Revocation;
 pub use tag::Tag;
